fix(inspector): add copy button to user messages
Extract CopyButton into a shared component and add it to UserMessage,
matching the existing assistant message UX.

Resolves #1555
